backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
twilio==8.10.0
pillow==10.1.0
aiofiles==23.2.1
python-socketio==5.11.0
schedule==1.2.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
email-validator==2.1.0